Consolidate placement rules based on cluster selectors
In order to know what to call the common placement rule, the new fields of
policyDefaults.placement.name and policies[0].placement.name are added.
If neither of these are set, then no placement rule consolidation will
occur.

Additionally, this now checks for duplicate placement rule names and
returns an error when that occurs.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>