Add support for patching manifests
This adds support for patching manifests that are wrapped in a policy.
This uses Kustomize internally, and thus the patches must be in the
format of Kustomize patches. An exception is that if the patches are
applied to only a single manifest, the apiVersion, kind, metadata.name,
and metadata.namespace fields are not necessary. They will be copied
over from the manifest before the patches are applied using Kustomize.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>