Add support for patching manifests (#11)
* Split out some of the logic in getPolicyTemplate

This improves the readability of the function. This will be especially
true after future changes.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>

* Create a helper function of unmarshalManifestBytes

This takes out functionality from unmarshalManifestFile and then
unmarshalManifestFile wraps unmarshalManifestBytes. This is required for
a future change.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>

* Add support for patching manifests

This adds support for patching manifests that are wrapped in a policy.
This uses Kustomize internally, and thus the patches must be in the
format of Kustomize patches. An exception is that if the patches are
applied to only a single manifest, the apiVersion, kind, metadata.name,
and metadata.namespace fields are not necessary. They will be copied
over from the manifest before the patches are applied using Kustomize.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>

* Fix the flakey TestGenerateSeparateBindings test

This stems from ranging of a map does not guarantee alphabetical order.
This change makes it so.

Signed-off-by: mprahl <mprahl@users.noreply.github.com>