Fail generation and add help on field typos
When the input configuration does not match the schema in the go type,
the generator will now respond with an error. The intention is to help
users when there is a typo in the configuration, which otherwise would
just be ignored (and in the case of a map, any configuration "inside"
would be dropped).

This also adds help text to these error cases, to try and identify what
field the user might have meant. That implementation uses a library
which does not compute the Levenshtein distance, but should work well
enough, and has the advantage of already being in the import graph.

Refs:
 - https://github.com/stolostron/backlog/issues/19646

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>