Add options for policy ordering
The new `dependencies` field in the generator is the same as the new
`dependencies` field in the Policy type, however, more fields are
optional and will be filled in by the generator with useful defaults.

The new `ignorePending` field is also the same, but since that would be
set on templates, and the generator does not expose much at the template
level, here the field is set on the policy, and the generator sets the
correct field in all the templates of that policy. For a similar reason,
the `extraDependencies` field is not exposed in the generator at all.

The new `orderViaDependencies` field will set up a chain of dependencies
on the created policies. So if 3 policies are created by the generator,
policy 2 will depend on policy 1, and policy 3 will depend on policy 2.

In all cases, dependency lists are merged - setting it on a specific
policy will *not* override the defaults, or a dependency added by the
`orderViaDependencies` flag.

Refs:
 - https://github.com/stolostron/backlog/issues/26183

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>