Add policy ordering tests
I wanted (basically) end to end tests for the ordering, since the
settings can affect multiple policies and manifests. These tests take
the generator config inline as YAML, and compare the output of the
plugin with a desired file.

It's slightly repetitive, but I think it's easier to read and write.

Refs:
 - https://github.com/stolostron/backlog/issues/26183

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>