Remove namespace from orderManifests dependencies
The namespace being added to these was incorrect - it should be empty
since the "correct" namespace on the managed cluster is the managed
cluster namespace... which has different names on different clusters.

Note: the dependencies from orderPolicies *are correct* - those are
always root policies, so the namespace is prepended to the name during
canonicalization.

This PR was going to make this work, but was decided against:
https://github.com/open-cluster-management-io/governance-policy-framework-addon/pull/15

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>